requirements.txt
streamlit
yt-dlp
openai-whisper
ffmpeg-python
openai
google-genai
python-dotenv
click
pillow
requests
